\name{lfstat_intro}
\alias{lfstat_intro}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Lfstat - Introduction
}
\description{
Opens a introduction-pdf-file
}
\usage{
lfstat_intro() 
}

\references{
Gustard, A. & Demuth, S. (2009) (Eds) Manual on Low-flow Estimation and Prediction. Operational Hydrology Report No. 50, WMO-No. 1029, 136p.
}

\author{
Daniel Koffler and Gregor Laaha
}

