\name{RcmdrPlugin.lfstat-package}
\alias{RcmdrPlugin.lfstat-package}
\alias{RcmdrPlugin.lfstat}
\docType{package}
\title{
\packageTitle{RcmdrPlugin.lfstat}
}
\description{
\packageDescription{RcmdrPlugin.lfstat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RcmdrPlugin.lfstat}
\packageIndices{RcmdrPlugin.lfstat}
}
\author{
\packageAuthor{RcmdrPlugin.lfstat}

Maintainer: \packageMaintainer{RcmdrPlugin.lfstat}
}
\keyword{ package }

