\name{RcmdrPlugin.lfstat-package}
\alias{RcmdrPlugin.lfstat-package}
\alias{RcmdrPlugin.lfstat}
\docType{package}
\title{
Install the lfstat Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" based on the lfstat package
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.lfstat\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-11-27\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
Daniel Koffler <daniel.koffler@boku.ac.at> and Gregor Laaha <gregor.laaha@boku.ac.at>

Maintainer: Daniel Koffler <daniel.koffler@boku.ac.at>
}

\keyword{ package }


